Add CardWidget_t, move card rendering out of client.c
CardWidget_t extends UIWidget_t with card state (text, colors, flags,
font). Card back pattern data, draw_card_back_pattern, and all front-face
rendering logic move from client.c into widgets/card.c. The render loop
now calls ui_widget_render() like every other widget. hovered/selected
live in UIWidget_t base; my_card and font are stored on the widget so no
extra params are needed at render time.
